chore(EarlySecurityChecksLLD): move hook inside specific folder for ESC
