+++ apps/ledger-live-desktop/src/renderer/components/SyncOnboarding/Manual/EarlySecurityChecks/useChangeLanguagePrompt.tsx
